Update for CrOS v70+
Check if writable container and debug flags are in "/usr/share/arc-setup/config.json"
Substitute '|mount rootfs rootfs / remount bind rw' for '|mount rootfs rootfs / remount bind ro' in Android's init.rc

+++ RootandSEpatch.sh
@@ -54,11 +54,30 @@ esac
 
 modify_cros_files() {
   
-# Just changing two environment variables for Android in /etc/init here.
-# Recent versions of CrOS have Android envs in arc-setup-env.
-# Older versions had envs in the .conf files
+# Just changing two environment variables for Android here.
+# In CrOS v70, the writeable container and debug flags have moved (again); they are now in "/usr/share/arc-setup/config.json"
+# (Older versions of CrOS have/had Android envs in arc-setup-env).
+# (Even older versions had envs in the .conf files.)
 
 mkdir -p /usr/local/Backup
+
+# As of CrOS v70, we need to modify the two values in /usr/share/arc-setup/config.json
+if [ -e /usr/share/arc-setup/config.json ]; then
+
+  mkdir -p /usr/local/Backup/arc-setup
+  echo "Copying usr/share/arc-setup/config.json to /usr/local/Backup/arc-setup/config.json.old"
+
+  cp -a /usr/share/arc-setup/config.json /usr/local/Backup/arc-setup/config.json.old
+  cp -a /usr/share/arc-setup/config.json /usr/share/arc-setup/config.json.old
+
+  echo "Setting '"ANDROID_DEBUGGABLE": true' and '"WRITABLE_MOUNT": true' in /usr/share/arc-setup/config.json"
+
+  sed -i 's/"ANDROID_DEBUGGABLE": false/"ANDROID_DEBUGGABLE": true/g' /usr/share/arc-setup/config.json 2>/dev/null
+  sed -i 's/"WRITABLE_MOUNT": false/"WRITABLE_MOUNT": true/g' /usr/share/arc-setup/config.json 2>/dev/null
+
+fi
+
+# In CrOS v6x, the two flags we want to change are within /etc/init/arc-setup-env
 
 if [ -e /etc/init/arc-setup-env ]; then
   echo "Copying /etc/init/arc-setup-env to /usr/local/Backup"
@@ -69,14 +88,20 @@ if [ -e /etc/init/arc-setup-env ]; then
   
   sed -i 's/export WRITABLE_MOUNT=0/export WRITABLE_MOUNT=1/g' /etc/init/arc-setup-env 2>/dev/null
   sed -i 's/export ANDROID_DEBUGGABLE=0/export ANDROID_DEBUGGABLE=1/g' /etc/init/arc-setup-env 2>/dev/null
+  
+# NOTE The below line (disabling shared fonts) is no longer needed as of recent CrOS versions.
+
   sed -i 's/export SHARE_FONTS=1/export SHARE_FONTS=0/g' /etc/init/arc-setup-env 2>/dev/null
-
-else
+  
+ fi
+ 
+ # In case we are running a really old version of CrOS somehow, support the original method:
+ 
+ if [ -e /usr/share/arc-setup/config.json ] || [ -e /etc/init/arc-setup-env ]
+ 
   echo "Copying /etc/init/arc-setup.conf and /etc/init/arc-system-mount.conf to /usr/local/Backup"
-
   sleep 0.2
-
-  echo "Setting 'env WRITABLE_MOUNT=1' in /etc/init/arc-setup.conf and /etc/init/arc-system-mount.conf"
+  echo "Setting 'env WRITABLE_MOUNT=1' in /etc/init/arc-setup.conf and/or /etc/init/arc-system-mount.conf"
 
   cp -a /etc/init/arc-system-mount.conf /usr/local/Backup/arc-system-mount.conf.old
   cp -a /etc/init/arc-system-mount.conf /etc/init/arc-system-mount.conf.old
@@ -1042,6 +1067,13 @@ echo "Adding 'import /init.super.rc' to existing init.rc"
 sed -i '7iimport /init.super.rc' $arc_system/../init.rc
 
 # SuperSU copying script ends
+
+# In recent CrOS versions (v70+), for a writable /system, currently it also seems to be necessary to edit init.rc (to switch 'ro' to 'rw')
+
+echo "Substituting '|mount rootfs rootfs / remount bind rw' for '|mount rootfs rootfs / remount bind ro' in existing init.rc"
+echo "A backup of init.rc will be stored as init.rc.old"
+
+sed -i.old 's|mount rootfs rootfs / remount bind ro|mount rootfs rootfs / remount bind rw|g' $arc_system/../init.rc
 
 copy_busybox
